Setup dealing damage over time to flying enemies
Harpoon tower now damages flying enemies over time when the harpoon is connected to the enemy.

+++ ProjectDefender/Assets/Scripts/Enemy/EnemyHeavy.cs
@@ -2,8 +2,8 @@ using UnityEngine;
 
 public class EnemyHeavy : Enemy
 {
-    [Header("Enemy Details")] [SerializeField]
-    private int shieldAmount = 50;
+    [Header("Enemy Details")] 
+    [SerializeField] private float shieldAmount = 50;
     [SerializeField] private EnemyShield shieldObject;
 
     protected override void Start()
@@ -20,7 +20,7 @@ public class EnemyHeavy : Enemy
         }
     }
 
-    public override void TakeDamage(int damage)
+    public override void TakeDamage(float damage)
     {
         if (shieldAmount > 0)
         {
